+++ site/homepage/content/release-notes.md
@@ -13,6 +13,7 @@ title = "Release Notes"
 * A new Quarkus based HTTP protocol adapter is now available. This adapter version provides better memory consumption and
   startup times comparing to the existing one.
 * The LoraWAN protocol adapter has been extended with support for the *Actility Enterprise* provider.
+* The LoraWAN protocol adapter has been extended with support for the *Orbiwise* provider.
 
 ## 1.3.0